+++ core/src/app/command/mod.rs
@@ -9,6 +9,7 @@ mod runnable;
 use super::AppMeta;
 use crate::reference::ReferenceCommand;
 use crate::storage::StorageCommand;
+use crate::time::TimeCommand;
 use crate::world::WorldCommand;
 use async_trait::async_trait;
 
@@ -18,6 +19,7 @@ pub enum Command {
     App(AppCommand),
     Reference(ReferenceCommand),
     Storage(StorageCommand),
+    Time(TimeCommand),
     World(WorldCommand),
 }
 
@@ -33,6 +35,7 @@ impl Runnable for Command {
             Self::App(c) => c.run(app_meta).await,
             Self::Reference(c) => c.run(app_meta).await,
             Self::Storage(c) => c.run(app_meta).await,
+            Self::Time(c) => c.run(app_meta).await,
             Self::World(c) => c.run(app_meta).await,
         }
     }
@@ -60,6 +63,11 @@ impl Runnable for Command {
                     .map(|c| c.into()),
             )
             .chain(
+                TimeCommand::parse_input(input, app_meta)
+                    .drain(..)
+                    .map(|c| c.into()),
+            )
+            .chain(
                 WorldCommand::parse_input(input, app_meta)
                     .drain(..)
                     .map(|c| c.into()),
@@ -73,6 +81,7 @@ impl Runnable for Command {
             .chain(AppCommand::autocomplete(input, app_meta).drain(..))
             .chain(ReferenceCommand::autocomplete(input, app_meta).drain(..))
             .chain(StorageCommand::autocomplete(input, app_meta).drain(..))
+            .chain(TimeCommand::autocomplete(input, app_meta).drain(..))
             .chain(WorldCommand::autocomplete(input, app_meta).drain(..))
             .collect();
 
@@ -104,6 +113,12 @@ impl From<ReferenceCommand> for Command {
 impl From<StorageCommand> for Command {
     fn from(c: StorageCommand) -> Command {
         Command::Storage(c)
+    }
+}
+
+impl From<TimeCommand> for Command {
+    fn from(c: TimeCommand) -> Command {
+        Command::Time(c)
     }
 }
 
@@ -154,6 +169,7 @@ mod test {
     fn autocomplete_test() {
         assert_eq!(
             [
+                ("date", "get the current time"),
                 ("delete [name]", "remove an entry from journal"),
                 ("dragonborn", "generate NPC species"),
                 ("druidic foci", "SRD item category"),