+++ core/src/lib.rs
@@ -7,6 +7,7 @@ pub use world::Thing;
 
 mod reference;
 mod storage;
+mod time;
 mod world;
 
 pub fn app(data_store: impl DataStore + 'static) -> app::App {